head_count_api/requirements.txt
Django==1.11.7
djangorestframework==3.8.2
psycopg2==2.7.3.2
pytz==2017.3
